Working with Scala 2.12.3\n\nDropped support for v2.3, v3.1 and v3.2

+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/PlanCacheMetricsMonitor.scala
@@ -29,7 +29,7 @@ class PlanCacheMetricsMonitor extends StringCacheMonitor {
 
   override def cacheDiscard(ignored1: String, ignored2: String, secondsSinceReplan: Int): Unit = {
     counter.incrementAndGet()
-    waitTime.addAndGet(secondsSinceReplan)
+    waitTime.addAndGet(secondsSinceReplan.toLong)
   }
 
   def numberOfReplans: Long = counter.get()
