+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_3/runtime/helpers/ValueConversion.scala
@@ -60,7 +60,7 @@ object ValueConversion {
     case s: String => Values.stringValue(s)
     case c: Char => Values.charValue(c)
     case d: Double => Values.doubleValue(d)
-    case f: Float => Values.doubleValue(f)
+    case f: Float => Values.doubleValue(f.toDouble)
     case n: Number => Values.longValue(n.longValue())
     case b: Boolean => Values.booleanValue(b)
     case n: Node => ValueUtils.fromNodeProxy(n)
